pyproject.toml
[tool.poetry]
name = "concierge-nlp-cluster"
version = "0.1.0"
description = ""
authors = ["felipe.manrique <admin@conciergebot.app>"]
readme = "README.md"
packages = [{include = "concierge_app"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = {extras = ["all"], version = "^0.95.1"}
uvicorn = "^0.21.1"
black = "^23.3.0"
pytest = "^7.3.0"
requests = "^2.28.2"
pypdf2 = "^3.0.1"
langchain = "^0.0.139"
pandas = "^2.0.0"
numpy = "^1.24.2"
python-multipart = "^0.0.6"
tiktoken = "^0.3.3"
python-docx = "^0.8.11"
python-dotenv = "^0.19.2"
chardet = "^5.1.0"
redis = "^4.5.4"
scipy = "^1.10.1"
langdetect = "^1.0.9"
pyyaml = "^6.0"
colorlog = "^6.7.0"
docx2txt = "^0.8"
openai = "^0.27.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
